Compute bounds from estimated components.

Compute bounds from estimated components.

Usage

post_instrument_bounds(
  gamma,
  sigma = 0,
  n.sigma = 5,
  estimates,
  df,
  outcome,
  post.inst,
  whichCI = "percentile",
  remove.outliers = T,
  conf = 0.95
)

Arguments

gamma

Integer or vector of integers giving lowest and highest value for mean causal effect of post-instrument variable M on outcome Y

sigma

Integer or vector of integers giving lowest and highest value for variance of causal effect of M on Y. Must be non-negative. Defaults to zero.

n.sigma

Integer: Number of distinct values of sigma (within bounds provided by sigma argument) for which to compute bounds. Defaults to 5.

estimates

An object from calling estimate_sensitivity() that contains a vector of point estimates t as well as bootstrap replicates t0

df

Data frame

outcome

Name of dependent variables of interest Y, a string

post.inst

Name of post-instrument variable M, a string

whichCI

Determines the type of boostrap confidence interval: percentile (the default), basic, or both

remove.outliers

Logical. If TRUE (the default), outliers from bootstrapped sampling distribution of the 2SLS regressions for Y and M are removed. The resulting sampling distribution should better approximate results from standard robust standard errors.

conf

The confidence level of the confidence intervals. Defaults to 0.95

Value

A matrix with values of gamma, sigma, point estimates or bounds of treatment effect, and confidence intervals for those point estimates or bounds.

Details

The function assumes that the instrument Z, treatment D, and post-instrument variable M are binary (in which case there is only the sensitivity parameter gamma and sigma can be left at its default value zero), or that z is binary and D and M continuous, in which case both gamma and sigma are sensitivity parameters.

References

Schuessler, J., Glynn, A. N., & Rueda, M.R. (2023). Post-Instrument Bias.

Glynn, A. N., Rueda, M. R., & Schuessler, J. (2023). Post-Instrument Bias in Linear Models. Sociological Methods & Research.

Examples

if (FALSE) {
gamma <- c(-0.02, 0.02)
sigma <- c(0, 0.002)
bounds <- post.instrument.bounds(gamma = gamma, sigma = sigma,
n.sigma = 2, estimates = out)}